+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # v3.0
 ## Additions
-* Added `ContinuousSpace` as a space option.
+* Added `ContinuousSpace` as a space option. Supports Euclidean and Cityblock metrics.
 * Universal plotting function `plotabm` that works for models with any kind of space.
 * new function `space_neighbors`, which works for any space. It always and consistently returns the **IDs** of neighbors irrespectively of the spatial structure.
 * `AgentBasedModel` now allows you to pass in an `AbstractAgent` type, or an instance of your agent.
@@ -13,6 +13,7 @@
 * Version of `add_agent!` now has keyword propagation as well (in case you make your types with `@kwdef` or Parameters.jl)
 * New function `nextid`
 * Cool new logo
+* `node_neighbors` now accepts a `neighbor_type` keyword for working with directed graphs
 
 ## Breaking Changes
 * Deprecated `Space` in favor of the individual spaces: `Nothing, GridSpace, GraphSpace, ContinuousSpace`.